Break safir.fastapi up into internal modules
Move the code for safir.fastapi out of __init__.py into internal
modules and turn __init__.py into only imports and re-exports.
